update go mod declaration
metrictank (which doesn't use go mod) imports this as
grafana/configparser.  but MT's lib gets used by carbon-relay-ng which
uses modules, and there we run into the problem that go.mod's
declaration doesn't match the repo. and this can't be fixed by a replace
directive. but this should do it.